+++ CHANGELOG.md
@@ -32,7 +32,7 @@ Some final annotations contained two of them: one from RATT and one from Prokka.
 * Fixed issues that occur when sequence IDs contain "|" character (#62).
 
 ### Housekeeping
-* Changed default generic ORF prefix to "HYBRAN" for greater clarity.
+* Changed default generic ORF prefix to "HYBRA" for greater clarity.
 * set default output directory to current directory.
 * Now ignoring warnings when generating translations for pseudogenes that aren't multiples of three.